[project][p2p] Remove dead p2p code, fold all p2p code into one file
