+++ api/service/discovery/service.go
@@ -10,7 +10,6 @@ import (
 	nodeconfig "github.com/harmony-one/harmony/internal/configs/node"
 	"github.com/harmony-one/harmony/internal/utils"
 	"github.com/harmony-one/harmony/p2p"
-	"github.com/harmony-one/harmony/p2p/host"
 )
 
 // Service is the struct for discovery service.
@@ -82,7 +81,7 @@ func (s *Service) contactP2pPeers() {
 	}
 	pingMsg := proto_discovery.NewPingMessage(s.host.GetSelfPeer(), s.config.IsClient)
 
-	msgBuf := host.ConstructP2pMessage(byte(0), pingMsg.ConstructPingMessage())
+	msgBuf := p2p.ConstructP2pMessage(byte(0), pingMsg.ConstructPingMessage())
 	s.sentPingMessage(s.config.ShardGroupID, msgBuf)
 
 	pingInterval := 5
