--- p2p/host/message.go
@@ -1,14 +1,0 @@
-package host
-
-import (
-	"encoding/binary"
-)
-
-// ConstructP2pMessage constructs the p2p message as [messageType, contentSize, content]
-func ConstructP2pMessage(msgType byte, content []byte) []byte {
-	message := make([]byte, 5+len(content))
-	message[0] = 17 // messageType 0x11
-	binary.BigEndian.PutUint32(message[1:5], uint32(len(content)))
-	copy(message[5:], content)
-	return message
-}
